Factor out manifest utilities
